+++ dart/lib/src/protocol/sentry_device.dart
@@ -19,6 +19,8 @@ class SentryDevice {
     this.manufacturer,
     this.brand,
     this.screenResolution,
+    this.screenHeightPixels,
+    this.screenWidthPixels,
     this.screenDensity,
     this.screenDpi,
     this.online,
@@ -34,6 +36,7 @@ class SentryDevice {
     this.externalFreeStorage,
     this.bootTime,
     this.timezone,
+    this.language,
   }) : assert(
           batteryLevel == null || (batteryLevel >= 0 && batteryLevel <= 100),
         );
@@ -49,6 +52,8 @@ class SentryDevice {
         manufacturer: data['manufacturer'],
         brand: data['brand'],
         screenResolution: data['screen_resolution'],
+        screenHeightPixels: data['screen_height_pixels'],
+        screenWidthPixels: data['screen_width_pixels'],
         screenDensity: data['screen_density'],
         screenDpi: data['screen_dpi'],
         online: data['online'],
@@ -66,6 +71,7 @@ class SentryDevice {
             ? DateTime.tryParse(data['boot_time'])
             : null,
         timezone: data['timezone'],
+        language: data['language'],
       );
 
   /// The name of the device. This is typically a hostname.
@@ -102,6 +108,12 @@ class SentryDevice {
 
   /// The screen resolution. (e.g.: `800x600`, `3040x1444`).
   final String? screenResolution;
+
+  /// The screen height in pixels. (e.g.: `600`, `1080`).
+  final int? screenHeightPixels;
+
+  /// The screen width in pixels. (e.g.: `800`, `1920`).
+  final int? screenWidthPixels;
 
   /// A floating point denoting the screen density.
   final double? screenDensity;
@@ -149,6 +161,9 @@ class SentryDevice {
 
   /// The timezone of the device, e.g.: `Europe/Vienna`.
   final String? timezone;
+
+  /// The language of the device, e.g.: `en_US`.
+  final String? language;
 
   /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
@@ -206,6 +221,14 @@ class SentryDevice {
 
     if (screenResolution != null) {
       json['screen_resolution'] = screenResolution;
+    }
+
+    if (screenWidthPixels != null) {
+      json['screen_width_pixels'] = screenWidthPixels;
+    }
+
+    if (screenHeightPixels != null) {
+      json['screen_height_pixels'] = screenHeightPixels;
     }
 
     if (screenDensity != null) {
@@ -268,6 +291,10 @@ class SentryDevice {
       json['timezone'] = timezone;
     }
 
+    if (language != null) {
+      json['language'] = language;
+    }
+
     return json;
   }
 
@@ -310,6 +337,8 @@ class SentryDevice {
     String? manufacturer,
     String? brand,
     String? screenResolution,
+    int? screenHeightPixels,
+    int? screenWidthPixels,
     double? screenDensity,
     int? screenDpi,
     bool? online,
@@ -325,6 +354,7 @@ class SentryDevice {
     int? externalFreeStorage,
     DateTime? bootTime,
     String? timezone,
+    String? language,
   }) =>
       SentryDevice(
         name: name ?? this.name,
@@ -337,6 +367,8 @@ class SentryDevice {
         manufacturer: manufacturer ?? this.manufacturer,
         brand: brand ?? this.brand,
         screenResolution: screenResolution ?? this.screenResolution,
+        screenHeightPixels: screenHeightPixels ?? this.screenHeightPixels,
+        screenWidthPixels: screenWidthPixels ?? this.screenWidthPixels,
         screenDensity: screenDensity ?? this.screenDensity,
         screenDpi: screenDpi ?? this.screenDpi,
         online: online ?? this.online,
@@ -352,5 +384,6 @@ class SentryDevice {
         externalFreeStorage: externalFreeStorage ?? this.externalFreeStorage,
         bootTime: bootTime ?? this.bootTime,
         timezone: timezone ?? this.timezone,
+        language: language ?? this.language,
       );
 }