+++ dart/lib/src/sentry.dart
@@ -138,19 +138,27 @@ class Sentry {
     SentryEvent event, {
     dynamic stackTrace,
     dynamic hint,
-  }) async =>
-      _hub.captureEvent(event, stackTrace: stackTrace, hint: hint);
+    ScopeCallback? withScope,
+  }) =>
+      _hub.captureEvent(
+        event,
+        stackTrace: stackTrace,
+        hint: hint,
+        withScope: withScope,
+      );
 
   /// Reports the [throwable] and optionally its [stackTrace] to Sentry.io.
   static Future<SentryId> captureException(
     dynamic throwable, {
     dynamic stackTrace,
     dynamic hint,
-  }) async =>
+    ScopeCallback? withScope,
+  }) =>
       _hub.captureException(
         throwable,
         stackTrace: stackTrace,
         hint: hint,
+        withScope: withScope,
       );
 
   static Future<SentryId> captureMessage(
@@ -159,13 +167,15 @@ class Sentry {
     String? template,
     List<dynamic>? params,
     dynamic hint,
-  }) async =>
+    ScopeCallback? withScope,
+  }) =>
       _hub.captureMessage(
         message,
         level: level,
         template: template,
         params: params,
         hint: hint,
+        withScope: withScope,
       );
 
   /// Close the client SDK