+++ dart/lib/src/sentry_options.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:developer';
 
 import 'package:http/http.dart';
 
@@ -158,6 +159,10 @@ class SentryOptions {
   /// As a result, you will get new groups as you enable or disable this flag for certain events.
   bool attachStacktrace = true;
 
+  /// Enable this option if you want to record calls to `print()` as
+  /// breadcrumbs.
+  bool enablePrintBreadcrumbs = true;
+
   /// If [platformChecker] is provided, it is used get the envirnoment.
   /// This is useful in tests. Should be an implementation of [PlatformChecker].
   PlatformChecker platformChecker = PlatformChecker();
@@ -250,5 +255,10 @@ void noOpLogger(SentryLevel level, String message) {}
 
 /// A Logger that prints out the level and message
 void dartLogger(SentryLevel level, String message) {
-  print('[${level.name}] $message');
+  log(
+    '[${level.name}] $message',
+    level: level.toDartLogLevel(),
+    name: 'sentry',
+    time: getUtcDateTime(),
+  );
 }