+++ dart/test/hub_test.dart
@@ -165,4 +165,78 @@ void main() {
     // TODO I'm not sure how to test it
     // could we set [hub.stack] as @visibleForTesting ?
   });
+
+  group('Hub withScope', () {
+    late Fixture fixture;
+
+    setUp(() {
+      fixture = Fixture();
+    });
+
+    test('captureEvent should create a new scope', () async {
+      final hub = fixture.getSut();
+      await hub.captureEvent(SentryEvent());
+      await hub.captureEvent(SentryEvent(), withScope: (scope) {
+        scope.user = SentryUser(id: 'foo bar');
+      });
+      await hub.captureEvent(SentryEvent());
+
+      var calls = fixture.client.captureEventCalls;
+      expect(calls.length, 3);
+      expect(calls[0].scope?.user, isNull);
+      expect(calls[1].scope?.user?.id, 'foo bar');
+      expect(calls[2].scope?.user, isNull);
+    });
+
+    test('captureException should create a new scope', () async {
+      final hub = fixture.getSut();
+      await hub.captureException(Exception('0'));
+      await hub.captureException(Exception('1'), withScope: (scope) {
+        scope.user = SentryUser(id: 'foo bar');
+      });
+      await hub.captureException(Exception('2'));
+
+      var calls = fixture.client.captureExceptionCalls;
+      expect(calls.length, 3);
+      expect(calls[0].scope?.user, isNull);
+      expect(calls[0].throwable?.toString(), 'Exception: 0');
+
+      expect(calls[1].scope?.user?.id, 'foo bar');
+      expect(calls[1].throwable?.toString(), 'Exception: 1');
+
+      expect(calls[2].scope?.user, isNull);
+      expect(calls[2].throwable?.toString(), 'Exception: 2');
+    });
+
+    test('captureMessage should create a new scope', () async {
+      final hub = fixture.getSut();
+      await hub.captureMessage('foo bar 0');
+      await hub.captureMessage('foo bar 1', withScope: (scope) {
+        scope.user = SentryUser(id: 'foo bar');
+      });
+      await hub.captureMessage('foo bar 2');
+
+      var calls = fixture.client.captureMessageCalls;
+      expect(calls.length, 3);
+      expect(calls[0].scope?.user, isNull);
+      expect(calls[0].formatted, 'foo bar 0');
+
+      expect(calls[1].scope?.user?.id, 'foo bar');
+      expect(calls[1].formatted, 'foo bar 1');
+
+      expect(calls[2].scope?.user, isNull);
+      expect(calls[2].formatted, 'foo bar 2');
+    });
+  });
+}
+
+class Fixture {
+  final MockSentryClient client = MockSentryClient();
+  final SentryOptions options = SentryOptions(dsn: fakeDsn);
+
+  Hub getSut() {
+    final hub = Hub(options);
+    hub.bindClient(client);
+    return hub;
+  }
 }